move count badge to utils
